**Issue**

https://github.com/gravitee-io/issues/issues/7187

**Description**

feat: personal token matching performance improvment

In TokenServiceImpl, the process matching the incoming request token with the token stored in database is not optimized and consumes a lot of CPU, causing troubles to our customer.

It searches all tokens in tokens collection, and, for each token : bcrypt the input token in order to compare it to database token.

So, for example, if there are 100 tokens in database,
According to the database tokens ordering ; if the second token is matching, there will be 2 bcrypt operations.
But if the 99th token is matching, there will be 99 bcrypt operations, which is consuming lot of time and CPU.

With the current Bcrypt algorithm (using a random salt), there is no way to compare the input token with the database token without Bcrypting it.

The workaround is to prioritize the checks of the token that were used most recently.
In this way, if there are 2 successive calls with the same token, the 2nd call will check only 1 token.

**Benchmark**

**Scenario 1 :**
On local computer.
100 tokens in database, none have been used yet.
100 **successive** calls to MAPI, on GET /apis.

|| Before      | After |
| ----------- | ----------- | ----------- |
|bCrypt operations count| 5451 | 149 |
|bCrypt count explanation| =100*random(1->100) | =random(1->100)+99 |
|max CPU| 7% | 7% |
|Execution time| > 6 minutes | 13 seconds |

In the `after`, first call parse mutliple tokens according to database ordering, but all 99 subsequent calls check only the first token.

**Scenario 2 :**
On local computer.
100 tokens in database, none have been used yet.
100 **parrallel simultaneous** calls to MAPI, on GET /apis.

|| Before      | After |
| ----------- | ----------- | ----------- |
|bCrypt operations count| 2900 | SAME |
|bCrypt count explanation| =100*random(1->100) | SAME |
|max CPU| 100% | SAME |
|Execution time| 30 seconds | SAME |

Launch scenario 2 again (now, the use date of the token has been set by the first launch).

|| Before      | After |
| ----------- | ----------- | ----------- |
|bCrypt operations count| 2900 | 100 |
|bCrypt count explanation| =100*random(1->100) | 100 |
|max CPU| 100% | 60% |
|Execution time| 30 seconds | 1 second |